+++ scripts/variables.sh
@@ -23,3 +23,5 @@ restore_processes=""
 restore_process_strategy_option="@resurrect-strategy-"
 
 inline_strategy_token="->"
+
+save_command_strategy_option="@resurrect-save-command-strategy"